requirements.txt
googlemaps
nylas
rich
prettytable
python-dotenv
